+++ src/helpers/constants/Express.js
@@ -1,0 +1,1 @@
+exports.queryStringOrderValues = ["asc", "ascending", "desc", "descending"];