+++ src/helpers/functions/Express.js
@@ -1,5 +1,6 @@
 const { validationResult, matchedData } = require("express-validator");
 const Error = require("../../classes/Error");
+const { queryStringOrderValues } = require("../constants/Express");
 
 exports.checkRequestData = req => {
     const errors = validationResult(req);
@@ -13,4 +14,11 @@ exports.checkRequestData = req => {
         cookies: matchedData(req, { locations: ["cookies"] }),
         headers: matchedData(req, { locations: ["headers"] }),
     };
+};
+
+exports.getQueryStringOrderValues = () => JSON.parse(JSON.stringify(queryStringOrderValues));
+
+exports.getMongooseOrderValueFromQueryString = str => {
+    const ascendingStrings = ["asc", "ascending"];
+    return ascendingStrings.includes(str) ? 1 : -1;
 };